Extract kubernetes reconciliation to separate file
